% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverageEstimation.R
\name{CalcHostLevelCoverage}
\alias{CalcHostLevelCoverage}
\title{Estimated probabilities of missing links based on the host Level Coverage Deficit}
\usage{
CalcHostLevelCoverage(list)
}
\arguments{
\item{list}{Network List}
}
\value{
A network list, with 'C_defmatrix', a matrix of probabilities based on coverage deficit, and
'OverallChaoEst' an estimate of the overall coverage deficit of the network.
}
\description{
Calls CoverageEstimator() to calculate host-level coverage deficit, then divides this by the number
 of unobserved interactions of that host.
}
