% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassandRa-package.R
\docType{package}
\name{cassandRa-package}
\alias{cassandRa}
\alias{cassandRa-package}
\title{cassandRa: Finds Missing Links and Metric Confidence Intervals in Ecological Bipartite Networks}
\description{
Provides methods to deal with under sampling in ecological bipartite networks from Terry and Lewis (2020) Ecology \doi{10.1002/ecy.3047} Includes tools to fit a variety of statistical network models and sample coverage estimators to highlight most likely missing links. Also includes simple functions to resample from observed networks to generate confidence intervals for common ecological network metrics.
}
\author{
\strong{Maintainer}: Chris Terry \email{christerry3@btinternet.com} [copyright holder]

}
\keyword{internal}
