% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCI.R
\name{ComputeCI}
\alias{ComputeCI}
\title{Compute Basic Confidence Intervals}
\usage{
ComputeCI(df)
}
\arguments{
\item{df}{A data frame produced by \code{RarefyNetwork()}}
}
\value{
a dataframe detailing confidence intervals at each tested sample size
}
\description{
Compute Basic Confidence Intervals
}
\examples{
 \dontrun{
 X<-RarefyNetwork(Safariland, n_per_level = 100)
 PlotRarefaction(X)
 }
}
