% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateListObject.R
\name{CreateListObject}
\alias{CreateListObject}
\title{Generates a network list from a food web}
\usage{
CreateListObject(web)
}
\arguments{
\item{web}{in format specified by the bipartite package. Rows = focal layer, columns = response layer}
}
\value{
A network list for use with other functions in EcoLinkPredict package
}
\description{
Gets a network in the base bipartite package format into a list format. N.B. Throughout this package
 uses hosts to refer to the focal layer, and 'wasps' the response layer, although this could equally
  be 'plants' and 'pollinators'.
}
\examples{
\dontrun{
demolist<-CreateListObject(Safariland)
str(demolist)
}
}
