% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoLinkPredict-package.R
\docType{package}
\name{EcoLinkPredict}
\alias{EcoLinkPredict}
\alias{EcoLinkPredict-package}
\title{Predicting Missing Links}
\description{
Package to make assessments of the most likely location of missing links and to test the effect of
 sample rarefaction or extrapolation on the confidence of bipartite network metrics.
}
\details{
The models to predict missing links are discussed in XXX.

See vignette for discussion and examples.
}
