% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCentrality.R
\name{FitCentrality}
\alias{FitCentrality}
\title{Fit Centrality Model}
\usage{
FitCentrality(list, N_runs = 10, maxit = 10000,
  method = "Nelder-Mead", ExtraSettings = NULL)
}
\arguments{
\item{list}{Network List}

\item{N_runs}{Number of start points to try. The best (maximum likelihood) half will be used to construct the probability matrix}

\item{maxit}{Default = 10'000}

\item{method}{Passed to optim, default = 'Nelder-Mead'}

\item{ExtraSettings}{Other control settings to pass to optim()}
}
\value{
Network list with added 'C_par', best fitting parameters, C_ProbsMatrix, the probability matrix
}
\description{
Repeatedly fits a centrality model to a binary interaction network to return a probability matrix
}
