% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSBM.R
\name{FitSBM}
\alias{FitSBM}
\title{Fit SBM Model}
\usage{
FitSBM(list, n_SBM = 10, G = NULL)
}
\arguments{
\item{list}{Network List}

\item{n_SBM}{Number of SBM models to fit. Default is 10. The top half (rounding up) are retained
and averaged to produce a probability matrix.}

\item{G}{The number of groups to divide the top layer and the focal layer into.}
}
\value{
Network list with 'SBM_ProbsMat', a matrix of probabilities assigned to each possible interaction,
 'SBM1', the best model fit derived from Optimise_SBM(), and 'SBM_G', the number of fitted groups.
}
\description{
Fit SBM Model
}
