% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortResponseCategory.R
\name{SortResponseCategory}
\alias{SortResponseCategory}
\title{Adds a dataframe that defines each interaction as true positive, false negative or true negative}
\usage{
SortResponseCategory(list)
}
\arguments{
\item{list}{Network list}
}
\value{
A Network list object with ObsSuccess, a dataframe detailing all the interactions and whether they are True Positives, False Negative or True Negatives
}
\description{
Adds a dataframe that defines each interaction as true positive, false negative or true negative
}
