% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAllModels.R
\name{TestAllModels}
\alias{TestAllModels}
\title{Test the models by AUC}
\usage{
TestAllModels(list)
}
\arguments{
\item{list}{A network list}
}
\value{
the network list with added AUC data. Key values are 'AUC', a dataframe with the AUC of
 each model and many combinations.
}
\description{
The function assumes FitAllModels() has already been run. It is a wrapper for 'SortResponseCategory()' and 'TestAUC()'
}
