% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAUC.R
\name{TestAUC}
\alias{TestAUC}
\title{Test via AUC the predictive capacity of each model or combination of models}
\usage{
TestAUC(list)
}
\arguments{
\item{list}{Network List}
}
\value{
a list with 'DataforAUC', a data frame with each interaction as a row and the predictions
 of each model, and 'AUC', a data frame with the predictive capacity of all the models and many combinations
}
\description{
Test via AUC the predictive capacity of each model or combination of models
}
