% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitAllModels.R
\name{FitAllModels}
\alias{FitAllModels}
\title{Fit all the models}
\usage{
FitAllModels(list, RepeatModels = 10)
}
\arguments{
\item{list}{A network list}

\item{RepeatModels}{How many times to fit each model from different starting points. Uses best half (rounding up)}
}
\value{
A network list including the model fit
}
\description{
Internal function called by \code{PredictLinks()}
 Fits the coverage deficit, Trait, Centrality, Matching-Centrality and SBM models by sequentially
 calling the individual functions.
}
