% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRarefaction.R
\name{PlotRarefaction}
\alias{PlotRarefaction}
\title{Plot Metric Response To Network Rarefaction}
\usage{
PlotRarefaction(df)
}
\arguments{
\item{df}{A data frame produced by RarefyNetwork}
}
\value{
A ggplot
}
\description{
Used to plot the output from \code{RarefyNetwork()}. See vignette!
}
\examples{
 data(Safariland, package = 'bipartite')
 X<-RarefyNetwork(Safariland, n_per_level = 100)
 ComputeCI(X)

}
